# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Stitch API Gateway",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/compliance/mtrs": {
      "get": {
        "tags": [
          "Compliance"
        ],
        "summary": "List Mtrs",
        "operationId": "list_mtrs_api_v1_compliance_mtrs_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "issue_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by issue date",
              "title": "Issue Date"
            },
            "description": "Filter by issue date"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status",
              "title": "Status"
            },
            "description": "Filter by status"
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WasteManifestSchema"
                  },
                  "title": "Response List Mtrs Api V1 Compliance Mtrs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/compliance/mtrs/{service_order_id}/generate": {
      "post": {
        "tags": [
          "Compliance"
        ],
        "summary": "Generate Mtr For Order",
        "operationId": "generate_mtr_for_order_api_v1_compliance_mtrs__service_order_id__generate_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "service_order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Service Order Id"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WasteManifestSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WasteItemSchema": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "waste_type": {
            "type": "string",
            "title": "Waste Type"
          },
          "quantity": {
            "type": "string",
            "title": "Quantity"
          },
          "un_code": {
            "type": "string",
            "title": "Un Code"
          }
        },
        "type": "object",
        "required": [
          "id",
          "waste_type",
          "quantity",
          "un_code"
        ],
        "title": "WasteItemSchema"
      },
      "WasteManifestSchema": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "generator_company_id": {
            "type": "string",
            "format": "uuid",
            "title": "Generator Company Id"
          },
          "transporter_company_id": {
            "type": "string",
            "format": "uuid",
            "title": "Transporter Company Id"
          },
          "service_order_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Order Id"
          },
          "issue_date": {
            "type": "string",
            "format": "date",
            "title": "Issue Date"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "driver_name": {
            "type": "string",
            "title": "Driver Name"
          },
          "vehicle_plate": {
            "type": "string",
            "title": "Vehicle Plate"
          },
          "expiration_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiration Date"
          },
          "usage_type": {
            "type": "string",
            "title": "Usage Type",
            "default": "SINGLE_USE"
          },
          "current_usages": {
            "type": "integer",
            "title": "Current Usages",
            "default": 0
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/WasteItemSchema"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "generator_company_id",
          "transporter_company_id",
          "issue_date",
          "status",
          "driver_name",
          "vehicle_plate",
          "items"
        ],
        "title": "WasteManifestSchema"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/login"
          }
        }
      }
    }
  }
}